// src/enums/Types.ts
export enum Types {
  SetErrorMessage = 'SET_ERROR_MESSAGE',
  SetTodosToState = 'SET_TODOS_TO_STATE',
  SetUpdatedTodoId = 'SET_UPDATED_TODO_ID',
  RemoveUpdatedTodoId = 'REMOVE_UPDATED_TODO_ID',
  CreateTodo = 'CREATE_TODO',
  DeleteTodo = 'DELETE_TODO',
  ToggleCompletedTodo = 'TOGGLE_COMPLETED_TODO',
  ClearCompletedTodo = 'CLEAR_COMPLETED_TODO',
  FilterBy = 'FILTER_BY',
  ToggleSelectAllTodo = 'TOGGLE_SELECT_ALL_TODO',
  EditTodo = 'EDIT_TODO',
}